feat: Massive Refactor
Created heatmap_utils to reduce the size of main.py and data.py, as well as many other minor fixes
